requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cssbeautifier==1.15.1
Django==5.1.2
django-celery-beat==2.7.0
django-cors-headers==4.5.0
django-filter==24.3
django-rest-framework==0.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djlint==1.35.2
drf-yasg==1.21.8
EditorConfig==0.12.4
html-tag-names==0.1.2
html-void-elements==0.1.0
inflection==0.5.1
jsbeautifier==1.15.1
json5==0.9.25
kombu==5.4.2
Markdown==3.7
packaging==24.1
pathspec==0.12.1
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
regex==2024.9.11
ruff==0.7.1
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.5
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.1
